Refactor NS manager DC, and NS provisioning keystone
