+++ ns-provisioning/helpers/init.rb
@@ -24,6 +24,6 @@ require_relative 'pop'
 require_relative 'hot'
 require_relative 'utils'
 require_relative 'instantiation'
-require_relative 'authentication'
-require_relative 'keystone_v2'
-require_relative 'keystone_v3'
+require_relative 'keystone/authentication'
+require_relative 'keystone/keystone_v2'
+require_relative 'keystone/keystone_v3'
