+++ ns-provisioning/helpers/keystone/authentication.rb
