+++ ns-provisioning/helpers/keystone/keystone_v3.rb
